+++ src/main/kotlin/ui/instance/context/InstanceContextMenu.kt
@@ -3,6 +3,7 @@ package dev.nycode.omsilauncher.ui.instance.context
 import dev.nycode.omsilauncher.ui.instance.context.impl.ActivateInstanceContextMenuAction
 import dev.nycode.omsilauncher.ui.instance.context.impl.DeleteInstanceContextMenuAction
 import dev.nycode.omsilauncher.ui.instance.context.impl.EditInstanceContextMenuAction
+import dev.nycode.omsilauncher.ui.instance.context.impl.ShowInstanceFilesAction
 import dev.nycode.omsilauncher.ui.instance.context.impl.start.EditorStartInstanceContextMenuAction
 import dev.nycode.omsilauncher.ui.instance.context.impl.start.RegularStartInstanceContextMenuAction
 
@@ -12,5 +13,6 @@ val instanceContextMenus: List<InstanceContextMenuAction> =
         EditorStartInstanceContextMenuAction,
         ActivateInstanceContextMenuAction,
         EditInstanceContextMenuAction,
-        DeleteInstanceContextMenuAction
+        DeleteInstanceContextMenuAction,
+        ShowInstanceFilesAction
     )